# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/actions": {
      "post": {
        "description": "Create a reusable action template for workflows. Actions define API calls that can be executed within workflow steps.",
        "operationId": "ActionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Action template configuration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Human-readable action name",
                    "example": "Send Slack Message"
                  },
                  "key": {
                    "type": "string",
                    "description": "Unique identifier for the action",
                    "example": "slack_send_message"
                  },
                  "toolId": {
                    "type": "string",
                    "description": "ID of the associated tool",
                    "example": "tool_slack_123"
                  },
                  "method": {
                    "type": "string",
                    "description": "HTTP method for the action",
                    "enum": [
                      "GET",
                      "POST",
                      "PUT",
                      "DELETE",
                      "PATCH"
                    ],
                    "example": "POST"
                  },
                  "endpoint": {
                    "type": "string",
                    "description": "API endpoint URL or path",
                    "example": "/api/chat.postMessage"
                  },
                  "headers": {
                    "type": "object",
                    "description": "HTTP headers required for the action",
                    "example": {
                      "Content-Type": "application/json",
                      "Authorization": "Bearer {token}"
                    }
                  },
                  "inputSchema": {
                    "type": "array",
                    "description": "Schema defining input parameters for the action",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "channel"
                        },
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "required": {
                          "type": "boolean",
                          "example": true
                        },
                        "description": {
                          "type": "string",
                          "example": "Slack channel ID"
                        }
                      }
                    },
                    "example": [
                      {
                        "name": "channel",
                        "type": "string",
                        "required": true,
                        "description": "Slack channel ID"
                      },
                      {
                        "name": "text",
                        "type": "string",
                        "required": true,
                        "description": "Message content"
                      }
                    ]
                  },
                  "executeIf": {
                    "type": "object",
                    "description": "Conditional execution rules (optional)",
                    "example": {
                      "user.role": "admin"
                    }
                  },
                  "version": {
                    "type": "number",
                    "description": "Action version number",
                    "example": 1
                  }
                },
                "required": [
                  "name",
                  "key",
                  "toolId",
                  "method",
                  "endpoint",
                  "headers",
                  "inputSchema"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Action created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "action_abc123"
                    },
                    "name": {
                      "type": "string",
                      "example": "Send Slack Message"
                    },
                    "key": {
                      "type": "string",
                      "example": "slack_send_message"
                    },
                    "toolId": {
                      "type": "string",
                      "example": "tool_slack_123"
                    },
                    "method": {
                      "type": "string",
                      "example": "POST"
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "/api/chat.postMessage"
                    },
                    "headers": {
                      "type": "object",
                      "example": {
                        "Content-Type": "application/json"
                      }
                    },
                    "inputSchema": {
                      "type": "array",
                      "example": []
                    },
                    "executeIf": {
                      "type": "object",
                      "example": {}
                    },
                    "version": {
                      "type": "number",
                      "example": 1
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid action configuration"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Create Action",
        "tags": [
          "Actions"
        ]
      },
      "get": {
        "description": "Get all reusable action templates for the organization",
        "operationId": "ActionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved actions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "action_abc123"
                      },
                      "name": {
                        "type": "string",
                        "example": "Send Slack Message"
                      },
                      "key": {
                        "type": "string",
                        "example": "slack_send_message"
                      },
                      "toolId": {
                        "type": "string",
                        "example": "tool_slack_123"
                      },
                      "method": {
                        "type": "string",
                        "example": "POST"
                      },
                      "endpoint": {
                        "type": "string",
                        "example": "/api/chat.postMessage"
                      },
                      "headers": {
                        "type": "object",
                        "example": {
                          "Content-Type": "application/json"
                        }
                      },
                      "inputSchema": {
                        "type": "array",
                        "example": [
                          {
                            "name": "channel",
                            "type": "string",
                            "required": true,
                            "description": "Slack channel ID"
                          }
                        ]
                      },
                      "executeIf": {
                        "type": "object",
                        "example": {}
                      },
                      "version": {
                        "type": "number",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "List Actions",
        "tags": [
          "Actions"
        ]
      }
    },
    "/actions/{id}": {
      "get": {
        "description": "Get a specific action template by ID",
        "operationId": "ActionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID",
            "schema": {
              "example": "action_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved action",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "action_abc123"
                    },
                    "name": {
                      "type": "string",
                      "example": "Send Slack Message"
                    },
                    "key": {
                      "type": "string",
                      "example": "slack_send_message"
                    },
                    "toolId": {
                      "type": "string",
                      "example": "tool_slack_123"
                    },
                    "method": {
                      "type": "string",
                      "example": "POST"
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "/api/chat.postMessage"
                    },
                    "headers": {
                      "type": "object",
                      "example": {
                        "Content-Type": "application/json",
                        "Authorization": "Bearer {token}"
                      }
                    },
                    "inputSchema": {
                      "type": "array",
                      "example": [
                        {
                          "name": "channel",
                          "type": "string",
                          "required": true,
                          "description": "Slack channel ID"
                        },
                        {
                          "name": "text",
                          "type": "string",
                          "required": true,
                          "description": "Message content"
                        }
                      ]
                    },
                    "executeIf": {
                      "type": "object",
                      "example": {
                        "user.role": "admin"
                      }
                    },
                    "version": {
                      "type": "number",
                      "example": 1
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Action not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Get Action",
        "tags": [
          "Actions"
        ]
      },
      "put": {
        "description": "Update an action template configuration",
        "operationId": "ActionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID",
            "schema": {
              "example": "action_abc123",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated action configuration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Human-readable action name",
                    "example": "Send Enhanced Slack Message"
                  },
                  "key": {
                    "type": "string",
                    "description": "Unique identifier for the action",
                    "example": "slack_send_enhanced_message"
                  },
                  "toolId": {
                    "type": "string",
                    "description": "ID of the associated tool",
                    "example": "tool_slack_123"
                  },
                  "method": {
                    "type": "string",
                    "description": "HTTP method for the action",
                    "enum": [
                      "GET",
                      "POST",
                      "PUT",
                      "DELETE",
                      "PATCH"
                    ],
                    "example": "POST"
                  },
                  "endpoint": {
                    "type": "string",
                    "description": "API endpoint URL or path",
                    "example": "/api/chat.postMessage"
                  },
                  "headers": {
                    "type": "object",
                    "description": "HTTP headers required for the action",
                    "example": {
                      "Content-Type": "application/json",
                      "Authorization": "Bearer {token}"
                    }
                  },
                  "inputSchema": {
                    "type": "array",
                    "description": "Schema defining input parameters for the action",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "required": {
                          "type": "boolean"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    },
                    "example": [
                      {
                        "name": "channel",
                        "type": "string",
                        "required": true,
                        "description": "Slack channel ID"
                      },
                      {
                        "name": "text",
                        "type": "string",
                        "required": true,
                        "description": "Message content"
                      },
                      {
                        "name": "attachments",
                        "type": "array",
                        "required": false,
                        "description": "Message attachments"
                      }
                    ]
                  },
                  "executeIf": {
                    "type": "object",
                    "description": "Conditional execution rules (optional)",
                    "example": {
                      "user.role": "admin"
                    }
                  },
                  "version": {
                    "type": "number",
                    "description": "Action version number",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action updated successfully"
          },
          "400": {
            "description": "Invalid action configuration"
          },
          "404": {
            "description": "Action not found"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Update Action",
        "tags": [
          "Actions"
        ]
      },
      "delete": {
        "description": "Delete an action template permanently. This will affect any workflows that use this action.",
        "operationId": "ActionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Action ID",
            "schema": {
              "example": "action_abc123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Action deleted successfully"
          },
          "404": {
            "description": "Action not found"
          },
          "409": {
            "description": "Action is being used by active workflows and cannot be deleted"
          }
        },
        "security": [
          {
            "x-user-id": []
          },
          {
            "x-org-id": []
          }
        ],
        "summary": "Delete Action",
        "tags": [
          "Actions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tolstoy API",
    "description": "Interactive API reference for Tolstoy workflow automation platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tolstoy.getpullse.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id",
        "description": "Organization ID for multi-tenant access"
      },
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for request context"
      }
    },
...
  }
}